fix bugs in spmv, spmm call ; add new sparseMatrix traits ; fix bugs in simd::valid

+++ trunk/fflas-ffpack/fflas/fflas_simd/simd128_int16.inl
@@ -59,7 +59,7 @@ template <> struct Simd128_impl<true, true, true, 2> {
      * Check if the pointer p is a multiple of alignemnt
      */
     template<class T>
-    static constexpr bool valid(T* p) {return p%alignment == 0;}
+    static constexpr bool valid(T* p) {return (int64_t)p%alignment == 0;}
 
     /*
      * Check if the number n is a multiple of vect_size
